[Fix] LTX23ImageToVideoPipeline: cast image to VAE dtype before encode
The I2V engine warmup crashed with "Input type (float) and bias type
(c10::BFloat16) should be the same" because the image tensor was cast
to the latent dtype (float32) before `vae.encode`, while the VAE conv
weights are bf16. Cast to `self.vae.dtype` instead; init_latents are
re-cast to the latent dtype after encoding.

+++ vllm_omni/diffusion/models/ltx2/pipeline_ltx2_3.py
@@ -1653,7 +1653,8 @@ class LTX23ImageToVideoPipeline(LTX23Pipeline):
                 image = torch.stack(image, dim=0)
             else:
                 image = self.video_processor.preprocess(image, height=height, width=width)
-            image = image.to(device=device, dtype=dtype)
+            # Cast to VAE dtype here; init_latents are re-cast to `dtype` after encode.
+            image = image.to(device=device, dtype=self.vae.dtype)
 
         height = height // self.vae_spatial_compression_ratio
         width = width // self.vae_spatial_compression_ratio
